+++ dcerpc/transport/dgfwd.c
@@ -158,7 +158,7 @@ PRIVATE unsigned32 rpc__dg_fwd_pkt (
     {
         RPC_DBG_GPRINTF(
             ("(rpc__dg_fwd_pkt) fwd map function returned error (st=%08lx, ptype=%s) [%s]\n", 
-            st,
+            (unsigned long)st,
             rpc__dg_pkt_name(RPC_DG_HDR_INQ_PTYPE(hdrp)), 
             rpc__dg_act_seq_string(hdrp)));
         return (FWD_PKT_NOTDONE);
